attempting to render d3 graph
