+++ package.json
@@ -13,6 +13,7 @@
   "author": "Daniel Dolich, Emily Krebs, Heidi Kim, Ross Sarcona, Tommy Liang",
   "license": "MIT",
   "devDependencies": {
+    "@types/d3": "^6.3.0",
     "@types/react": "^17.0.1",
     "@types/react-router-dom": "^5.1.7",
     "concurrently": "^5.3.0",
